Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@6531 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ apple2/test/Hex2BinTest.java
@@ -14,7 +14,7 @@ public class Hex2BinTest extends TestCase
 
     public void testHexbyte()
     {
-        assertEquals("00",Hex2Bin.hexbyte(0));
+        assertEquals("00",Hex2Bin.hexbyte((byte)0));
     }
 
 }
